requirements.txt
asgiref==3.5.2
astpretty==3.0.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.6
robotframework==6.0.1
sqlparse==0.4.3
